+++ src/actions/authActions.ts
@@ -1,12 +1,27 @@
 "use server";
 
-import { sql } from "@/lib/db";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { commonHeaders } from "@/config/commonHeaders";
+import { AUTH_API_URL } from "@/constants/apiEndpoints";
+import { SignupForm } from "@/types/type";
+import { getServerSession } from "next-auth";
 
-export const checkUser = async (email: string) => {
-  const result = await sql`
-        SELECT COUNT(*)
-        FROM users
-        WHERE email = ${email}`;
+export const userSignUp = async (data: SignupForm) => {
+  try {
+    const res = await fetch(AUTH_API_URL.signUp, {
+      method: "POST",
+      headers: commonHeaders,
+      body: JSON.stringify(data),
+    });
 
-  return result[0].count > 0;
+    if (res.ok) {
+      const session = await getServerSession(authOptions);
+      const { data } = await res.json();
+      session.accessToken = data.accessToken;
+      return `/`;
+    }
+  } catch (error) {
+    console.error(error);
+    return false;
+  }
 };
