--- src/lib/db.ts
@@ -1,3 +1,0 @@
-import { neon } from "@neondatabase/serverless";
-
-export const sql = neon(process.env.DATABASE_URL!);
