--- src/services/postServices.ts
@@ -1,19 +1,0 @@
-export const fetchAllPosts = async () => {
-  try {
-    const response = undefined;
-    // const PJRes = await fetch(`${BASE_URL}/api/v1/project`);
-    // const PJData = await PJRes.json();
-    // const PRRes = await fetch(`${BASE_URL}/api/v1/pr`);
-    // const PRdata = await PRRes.json();
-    // // // const [PJres, PRres] = await Promise.all([PJreq, PRreq]);
-    // // const response = PJreq.data.data.teamRecruits;
-    // // //   [
-    // //   ...PJres.data.data.teamRecruits,
-    // //   ...PRres.data.data.content,
-    // // ];
-    return response;
-  } catch (error) {
-    console.error(error);
-    return;
-  }
-};
